requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
gspread==5.12.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
